chore: disable type checking for migrated scripts
